Improved cleanup for SQL queries.
Previously, queries with multiple bindings were generating multiple
metric entries. The queries below:

    select name from table where id in (?, ?)
    select name from table where id in (?, ?, ?)

were grouped as separate queries, which makes no sense.

Now both of them will be registered in a single metric as:
    select name from table where id in (?)

For insert queries:

    insert into table (name) values ('name1'), ('name2'), ('name3')
    insert into table (name) values ('name4')
    insert into table (name) values (?)

the data is not relevant, so they'll be registered as:

    insert into table (name) values ()

This simplifies our dashboard filters and the metrics become more
meaningful.