Fixes a problem with a inefficient regex and with changes in the config.
Changed a regex to avoid backtracking, making it faster and avoiding
problems with larger strings. And if any of the regexes fails, the "query"
label will contain "[error]" instead of breaking the application.

Changed how we deal with metrics when we change the environment variable
PROMETHEUS_COLLECT_FULL_SQL_QUERY. Previously it would create issues with
wrong number of labels if you changed this config and already had metrics
stored. Now if you disable the collecting of full SQL queries, the "query"
label will always have the value "[omitted]".